Add ResultComponent, guard, and directive
A ResultComponent with corresponding TypeScript, HTML, and CSS files has been added to handle the result feature. Alongside this, a `CanActivate` guard was introduced to protect routes, which requires the `id` parameter in the `result` route. A 'typing' directive is also included to animate typing effect in the application. The commit brings changes in the `app.routes.ts` by including routes for the ResultComponent and applying the newly added guard. The GenerateComponent file was updated to utilize the new directive and handle navigation to the `result` route. Inclusions of new packages `prettier` and `spectre.css` are noticeable in package files. Styling modifications are also made in relevant SCSS files.