Refactor layout and improve loader's visibility
The code modifies the layout of the generate component in both HTML and SCSS files, to improve the visibility and positioning of the loading screen. Changes include the introduction of a new loader__container class for center alignment, modification of existing element arrangements, and commenting out of some styles in position class. The typing-animation class has also been adjusted for center alignment. The refactoring ensures the loading screen takes up more space in the viewport and appears in the center, enhancing the user experience during content generation.