Adjust margins in result.component.scss file
Modifications were made to the result.component.scss file in the generate feature. A new line to add a margin-bottom to the .page class was introduced and the margin-top in #brief__overview and #brief__content was switched from px to rem. This ensures better consistency and modularity in the component's styling.