Add expand icon and related styling for inspiration image
An expand icon was added to the inspiration image along with related CSS styling. The icon SVG was added to the assets directory, and new classes were created in the SCSS file to style both the icon and the image. An inspiration section was added to the HTML file to display the image with the expand icon.