Update .gitignore to track specific environment files
The changes in this commit are primarily an update in the .gitignore file. Specifically, it adjusts the patterns to ignore environment files, instead favoring the tracking of only environment.ts and environment.development.ts. This helps to ensure that the necessary environment files are tracked, while keeping unnecessary ones out of the repository.