Add gh-pages package and update build script
The gh-pages package has been added to both package.json and package-lock.json files to aid in deployment. The build script in the .github/workflows/main.yml file has been updated, specifying 'dist' as the directory for the gh-pages command. This will allow for correct deployment of the built files.